Reverted GenericType implementation for inputs and prompts that are arrays. Decided to make inputs or prompts IList<string> by default.
Changed Stop on TextCompletionRequest to IList<string>

This is a beraking change if you were using the Stop string on TextCompletionRequest. All costructors still support string and IList<string>.

Extension methods re not affected by this change and continue to work as they did previously.

There is also a new extension method for converting a string to a Ilist<string>
e.g
"test".ToList()

Any new API's that support arrays or string will be implmented as IList<string> by default unless there is a specific reason not to do so.